Remove more Link UI.
