+++ link/api/link.api
@@ -286,24 +286,6 @@ public final class com/stripe/android/link/ui/inline/UserInput$SignUp : com/stri
 	public fun toString ()Ljava/lang/String;
 }
 
-public final class com/stripe/android/link/ui/signup/ComposableSingletons$SignUpScreenKt {
-	public static final field INSTANCE Lcom/stripe/android/link/ui/signup/ComposableSingletons$SignUpScreenKt;
-	public static field lambda-1 Lkotlin/jvm/functions/Function2;
-	public static field lambda-2 Lkotlin/jvm/functions/Function2;
-	public fun <init> ()V
-	public final fun getLambda-1$link_release ()Lkotlin/jvm/functions/Function2;
-	public final fun getLambda-2$link_release ()Lkotlin/jvm/functions/Function2;
-}
-
-public final class com/stripe/android/link/ui/verification/ComposableSingletons$VerificationScreenKt {
-	public static final field INSTANCE Lcom/stripe/android/link/ui/verification/ComposableSingletons$VerificationScreenKt;
-	public static field lambda-1 Lkotlin/jvm/functions/Function2;
-	public static field lambda-2 Lkotlin/jvm/functions/Function2;
-	public fun <init> ()V
-	public final fun getLambda-1$link_release ()Lkotlin/jvm/functions/Function2;
-	public final fun getLambda-2$link_release ()Lkotlin/jvm/functions/Function2;
-}
-
 public final class com/stripe/android/link/ui/verification/VerificationDialogKt {
 	public static final fun LinkVerificationDialog (Lcom/stripe/android/link/LinkPaymentLauncher;Lkotlin/jvm/functions/Function1;Landroidx/compose/runtime/Composer;I)V
 }
