+++ link/src/androidTest/java/com/stripe/android/link/AndroidComposeTestRule.kt
@@ -1,27 +1,7 @@
 package com.stripe.android.link
 
-import android.content.Context
-import android.content.Intent
 import androidx.activity.ComponentActivity
-import androidx.compose.ui.test.junit4.AndroidComposeTestRule
-import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.rules.ActivityScenarioRule
-
-/**
- * Factory method to provide android specific implementation of createComposeRule, for a given
- * activity class type A that needs to be launched via an intent.
- *
- * @param intentFactory A lambda that provides a Context that can used to create an intent. A intent needs to be returned.
- */
-inline fun <A : ComponentActivity> createAndroidIntentComposeRule(intentFactory: (context: Context) -> Intent): AndroidComposeTestRule<ActivityScenarioRule<A>, A> {
-    val context = ApplicationProvider.getApplicationContext<Context>()
-    val intent = intentFactory(context)
-
-    return AndroidComposeTestRule(
-        activityRule = ActivityScenarioRule(intent),
-        activityProvider = { scenarioRule -> scenarioRule.getActivity() }
-    )
-}
 
 /**
  * Gets the activity from a scenarioRule.
