+++ link/src/androidTest/java/com/stripe/android/link/ui/LinkAppBarStateTest.kt
@@ -20,7 +20,6 @@ internal class LinkAppBarStateTest {
     @Test
     fun rejectsBlankEmailAddress() {
         val state = buildLinkAppBarState(
-            isRootScreen = true,
             currentRoute = LinkScreen.Wallet.route,
             email = " ",
             accountStatus = AccountStatus.SignedOut
@@ -30,49 +29,8 @@ internal class LinkAppBarStateTest {
     }
 
     @Test
-    fun loadingScreenShowsCorrectAppBarState() {
-        val state = buildLinkAppBarState(
-            isRootScreen = true,
-            currentRoute = LinkScreen.Loading.route,
-            email = null,
-            accountStatus = AccountStatus.SignedOut
-        )
-
-        val expected = LinkAppBarState(
-            navigationIcon = R.drawable.stripe_link_close,
-            showHeader = true,
-            showOverflowMenu = false,
-            email = null,
-            accountStatus = AccountStatus.SignedOut
-        )
-
-        assertThat(state).isEqualTo(expected)
-    }
-
-    @Test
-    fun verificationScreenShowsCorrectAppBarState() {
-        val state = buildLinkAppBarState(
-            isRootScreen = true,
-            currentRoute = LinkScreen.Verification.route,
-            email = "someone@stripe.com",
-            accountStatus = AccountStatus.VerificationStarted
-        )
-
-        val expected = LinkAppBarState(
-            navigationIcon = R.drawable.stripe_link_close,
-            showHeader = true,
-            showOverflowMenu = false,
-            email = null,
-            accountStatus = AccountStatus.VerificationStarted
-        )
-
-        assertThat(state).isEqualTo(expected)
-    }
-
-    @Test
     fun verificationDialogShowsCorrectAppBarState() {
         val state = buildLinkAppBarState(
-            isRootScreen = true,
             currentRoute = LinkScreen.VerificationDialog.route,
             email = "someone@stripe.com",
             accountStatus = AccountStatus.VerificationStarted
@@ -80,7 +38,6 @@ internal class LinkAppBarStateTest {
 
         val expected = LinkAppBarState(
             navigationIcon = R.drawable.stripe_link_close,
-            showHeader = true,
             showOverflowMenu = false,
             email = "someone@stripe.com",
             accountStatus = AccountStatus.VerificationStarted
@@ -92,7 +49,6 @@ internal class LinkAppBarStateTest {
     @Test
     fun walletScreenShowsCorrectAppBarState() {
         val state = buildLinkAppBarState(
-            isRootScreen = true,
             currentRoute = LinkScreen.Wallet.route,
             email = "someone@stripe.com",
             accountStatus = AccountStatus.Verified
@@ -100,117 +56,15 @@ internal class LinkAppBarStateTest {
 
         val expected = LinkAppBarState(
             navigationIcon = R.drawable.stripe_link_close,
-            showHeader = true,
             showOverflowMenu = true,
             email = "someone@stripe.com",
             accountStatus = AccountStatus.Verified
-        )
-
-        assertThat(state).isEqualTo(expected)
-    }
-
-    @Test
-    fun paymentMethodScreenShowsCorrectAppBarStateWhenAddingFirstPaymentMethod() {
-        val state = buildLinkAppBarState(
-            isRootScreen = true,
-            currentRoute = LinkScreen.PaymentMethod.route,
-            email = "someone@stripe.com",
-            accountStatus = AccountStatus.Verified
-        )
-
-        val expected = LinkAppBarState(
-            navigationIcon = R.drawable.stripe_link_close,
-            showHeader = true,
-            showOverflowMenu = true,
-            email = "someone@stripe.com",
-            accountStatus = AccountStatus.Verified
-        )
-
-        assertThat(state).isEqualTo(expected)
-    }
-
-    @Test
-    fun paymentMethodScreenShowsCorrectAppBarStateWhenThereAreExistingPaymentMethods() {
-        val state = buildLinkAppBarState(
-            isRootScreen = false,
-            currentRoute = LinkScreen.PaymentMethod.route,
-            email = "someone@stripe.com",
-            accountStatus = AccountStatus.Verified
-        )
-
-        val expected = LinkAppBarState(
-            navigationIcon = R.drawable.stripe_link_back,
-            showHeader = false,
-            showOverflowMenu = false,
-            email = null,
-            accountStatus = AccountStatus.Verified
-        )
-
-        assertThat(state).isEqualTo(expected)
-    }
-
-    @Test
-    fun cardEditScreenShowsCorrectAppBarState() {
-        val state = buildLinkAppBarState(
-            isRootScreen = false,
-            currentRoute = LinkScreen.CardEdit.route,
-            email = "someone@stripe.com",
-            accountStatus = AccountStatus.Verified
-        )
-
-        val expected = LinkAppBarState(
-            navigationIcon = R.drawable.stripe_link_back,
-            showHeader = false,
-            showOverflowMenu = false,
-            email = null,
-            accountStatus = AccountStatus.Verified
-        )
-
-        assertThat(state).isEqualTo(expected)
-    }
-
-    @Test
-    fun signupScreenShowsCorrectAppBarState() {
-        val state = buildLinkAppBarState(
-            isRootScreen = true,
-            currentRoute = LinkScreen.SignUp.route,
-            email = null,
-            accountStatus = AccountStatus.SignedOut
-        )
-
-        val expected = LinkAppBarState(
-            navigationIcon = R.drawable.stripe_link_close,
-            showHeader = true,
-            showOverflowMenu = false,
-            email = null,
-            accountStatus = AccountStatus.SignedOut
-        )
-
-        assertThat(state).isEqualTo(expected)
-    }
-
-    @Test
-    fun signupScreenShowsCorrectAppBarStateWithEmail() {
-        val state = buildLinkAppBarState(
-            isRootScreen = true,
-            currentRoute = LinkScreen.SignUp.route,
-            email = "someone@stripe.com",
-            accountStatus = AccountStatus.NeedsVerification
-        )
-
-        val expected = LinkAppBarState(
-            navigationIcon = R.drawable.stripe_link_close,
-            showHeader = true,
-            showOverflowMenu = false,
-            email = null,
-            accountStatus = AccountStatus.NeedsVerification
         )
 
         assertThat(state).isEqualTo(expected)
     }
 
     private fun buildLinkAppBarState(
-        isRootScreen: Boolean,
         currentRoute: String?,
         email: String?,
         accountStatus: AccountStatus?
@@ -218,7 +72,7 @@ internal class LinkAppBarStateTest {
         var state: LinkAppBarState? = null
 
         composeTestRule.setContent {
-            state = rememberLinkAppBarState(isRootScreen, currentRoute, email, accountStatus)
+            state = rememberLinkAppBarState(true, currentRoute, email, accountStatus)
         }
 
         return state ?: throw AssertionError(
