+++ link/src/main/java/com/stripe/android/link/LinkPaymentLauncher.kt
@@ -21,7 +21,6 @@ import com.stripe.android.link.injection.LinkComponent
 import com.stripe.android.link.injection.LinkPaymentLauncherComponent
 import com.stripe.android.link.ui.inline.UserInput
 import com.stripe.android.link.ui.paymentmethod.SupportedPaymentMethod
-import com.stripe.android.link.ui.signup.SignUpViewModel
 import com.stripe.android.link.ui.verification.VerificationViewModel
 import com.stripe.android.model.PaymentMethodCreateParams
 import com.stripe.android.model.StripeIntent
@@ -195,7 +194,6 @@ class LinkPaymentLauncher @Inject internal constructor(
         val injector = object : NonFallbackInjector {
             override fun inject(injectable: Injectable<*>) {
                 when (injectable) {
-                    is SignUpViewModel.Factory -> linkComponent.inject(injectable)
                     is VerificationViewModel.Factory -> linkComponent.inject(injectable)
                     else -> {
                         throw IllegalArgumentException("invalid Injectable $injectable requested in $this")
