+++ link/src/main/java/com/stripe/android/link/ui/LinkAppBarState.kt
@@ -3,13 +3,11 @@ package com.stripe.android.link.ui
 import androidx.annotation.DrawableRes
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
-import com.stripe.android.link.LinkScreen
 import com.stripe.android.link.R
 import com.stripe.android.link.model.AccountStatus
 
 internal data class LinkAppBarState(
     @DrawableRes val navigationIcon: Int,
-    val showHeader: Boolean,
     val showOverflowMenu: Boolean,
     val email: String?,
     val accountStatus: AccountStatus?
@@ -23,20 +21,6 @@ internal fun rememberLinkAppBarState(
     accountStatus: AccountStatus?
 ): LinkAppBarState {
     return remember(currentRoute, email) {
-        val showHeader = when (currentRoute) {
-            LinkScreen.CardEdit.route -> false
-            LinkScreen.PaymentMethod.route -> isRootScreen
-            else -> true
-        }
-
-        val hideEmail = when (currentRoute) {
-            LinkScreen.CardEdit.route,
-            LinkScreen.Verification.route,
-            LinkScreen.SignUp.route -> true
-            LinkScreen.PaymentMethod.route -> !isRootScreen
-            else -> false
-        }
-
         // If there's an email address, we want to allow the user to log
         // out of the existing account.
         val showOverflowMenu = isRootScreen && email != null && accountStatus == AccountStatus.Verified
@@ -47,9 +31,8 @@ internal fun rememberLinkAppBarState(
             } else {
                 R.drawable.stripe_link_back
             },
-            showHeader = showHeader,
             showOverflowMenu = showOverflowMenu,
-            email = email?.takeUnless { it.isBlank() || hideEmail },
+            email = email?.takeUnless { it.isBlank() },
             accountStatus = accountStatus
         )
     }
