+++ link/src/main/java/com/stripe/android/link/ui/inline/InlineSignupViewModel.kt
@@ -14,7 +14,6 @@ import com.stripe.android.link.analytics.LinkEventsReporter
 import com.stripe.android.link.ui.ErrorMessage
 import com.stripe.android.link.ui.getErrorMessage
 import com.stripe.android.link.ui.signup.SignUpState
-import com.stripe.android.link.ui.signup.SignUpViewModel
 import com.stripe.android.model.PaymentIntent
 import com.stripe.android.model.SetupIntent
 import com.stripe.android.uicore.elements.EmailConfig
@@ -101,7 +100,7 @@ internal class InlineSignupViewModel @Inject constructor(
 
     private var hasExpanded = false
 
-    private var debouncer = SignUpViewModel.Debouncer(prefilledEmail)
+    private var debouncer = Debouncer(prefilledEmail)
 
     fun toggleExpanded() {
         _viewState.update { oldState ->
