+++ link/src/main/java/com/stripe/android/link/ui/verification/VerificationScreen.kt
@@ -6,16 +6,13 @@ import androidx.annotation.StringRes
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.material.ContentAlpha
 import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -33,69 +30,26 @@ import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.stripe.android.core.injection.NonFallbackInjector
 import com.stripe.android.link.R
 import com.stripe.android.link.model.LinkAccount
-import com.stripe.android.link.theme.DefaultLinkTheme
 import com.stripe.android.link.theme.linkColors
 import com.stripe.android.link.theme.linkShapes
 import com.stripe.android.link.ui.ErrorMessage
 import com.stripe.android.link.ui.ErrorText
 import com.stripe.android.link.ui.ScrollableTopLevelColumn
-import com.stripe.android.ui.core.elements.OTPSpec
 import com.stripe.android.uicore.DefaultStripeTheme
 import com.stripe.android.uicore.elements.OTPElement
 import com.stripe.android.uicore.elements.OTPElementUI
 import kotlinx.coroutines.delay
-
-@Preview
-@Composable
-private fun VerificationBodyPreview() {
-    DefaultLinkTheme {
-        Surface {
-            VerificationBody(
-                headerStringResId = R.string.stripe_verification_header,
-                messageStringResId = R.string.stripe_verification_message,
-                showChangeEmailMessage = true,
-                redactedPhoneNumber = "+1********23",
-                email = "test@stripe.com",
-                otpElement = OTPSpec.transform(),
-                isProcessing = false,
-                isSendingNewCode = false,
-                errorMessage = null,
-                focusRequester = remember { FocusRequester() },
-                onBack = { },
-                onChangeEmailClick = { },
-                onResendCodeClick = { }
-            )
-        }
-    }
-}
-
-@Composable
-internal fun VerificationBodyFullFlow(
-    linkAccount: LinkAccount,
-    injector: NonFallbackInjector
-) {
-    VerificationBody(
-        headerStringResId = R.string.stripe_verification_header,
-        messageStringResId = R.string.stripe_verification_message,
-        showChangeEmailMessage = true,
-        linkAccount = linkAccount,
-        injector = injector
-    )
-}
 
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
 internal fun VerificationBody(
     @StringRes headerStringResId: Int,
     @StringRes messageStringResId: Int,
-    showChangeEmailMessage: Boolean,
     linkAccount: LinkAccount,
     injector: NonFallbackInjector,
     onVerificationCompleted: (() -> Unit)? = null
@@ -146,16 +100,13 @@ internal fun VerificationBody(
     VerificationBody(
         headerStringResId = headerStringResId,
         messageStringResId = messageStringResId,
-        showChangeEmailMessage = showChangeEmailMessage,
         redactedPhoneNumber = viewModel.linkAccount.redactedPhoneNumber,
-        email = viewModel.linkAccount.email,
         otpElement = viewModel.otpElement,
         isProcessing = viewState.isProcessing,
         isSendingNewCode = viewState.isSendingNewCode,
         errorMessage = viewState.errorMessage,
         focusRequester = focusRequester,
         onBack = viewModel::onBack,
-        onChangeEmailClick = viewModel::onChangeEmailClicked,
         onResendCodeClick = viewModel::resendCode,
     )
 }
@@ -164,16 +115,13 @@ internal fun VerificationBody(
 internal fun VerificationBody(
     @StringRes headerStringResId: Int,
     @StringRes messageStringResId: Int,
-    showChangeEmailMessage: Boolean,
     redactedPhoneNumber: String,
-    email: String,
     otpElement: OTPElement,
     isProcessing: Boolean,
     isSendingNewCode: Boolean,
     errorMessage: ErrorMessage?,
     focusRequester: FocusRequester,
     onBack: () -> Unit,
-    onChangeEmailClick: () -> Unit,
     onResendCodeClick: () -> Unit
 ) {
     BackHandler(onBack = onBack)
@@ -206,14 +154,6 @@ internal fun VerificationBody(
             )
         }
 
-        if (showChangeEmailMessage) {
-            ChangeEmailRow(
-                email = email,
-                isProcessing = isProcessing,
-                onChangeEmailClick = onChangeEmailClick,
-            )
-        }
-
         AnimatedVisibility(visible = errorMessage != null) {
             ErrorText(
                 text = errorMessage?.getMessage(LocalContext.current.resources).orEmpty(),
@@ -225,39 +165,6 @@ internal fun VerificationBody(
             isProcessing = isProcessing,
             isSendingNewCode = isSendingNewCode,
             onClick = onResendCodeClick,
-        )
-    }
-}
-
-@Composable
-private fun ChangeEmailRow(
-    email: String,
-    isProcessing: Boolean,
-    onChangeEmailClick: () -> Unit,
-) {
-    Row(
-        modifier = Modifier.padding(vertical = 14.dp),
-        horizontalArrangement = Arrangement.Center
-    ) {
-        Text(
-            text = stringResource(id = R.string.stripe_verification_not_email, email),
-            modifier = Modifier.weight(weight = 1f, fill = false),
-            color = MaterialTheme.colors.onSecondary,
-            overflow = TextOverflow.Ellipsis,
-            maxLines = 1,
-            style = MaterialTheme.typography.body2
-        )
-        Text(
-            text = stringResource(id = R.string.stripe_verification_change_email),
-            modifier = Modifier
-                .padding(start = 4.dp)
-                .clickable(
-                    enabled = !isProcessing,
-                    onClick = onChangeEmailClick
-                ),
-            color = MaterialTheme.linkColors.actionLabel,
-            maxLines = 1,
-            style = MaterialTheme.typography.body2
         )
     }
 }
