+++ link/src/main/java/com/stripe/android/link/ui/verification/VerificationViewModel.kt
@@ -142,12 +142,6 @@ internal class VerificationViewModel @Inject constructor(
         linkAccountManager.logout()
     }
 
-    fun onChangeEmailClicked() {
-        clearError()
-        navigator.navigateTo(LinkScreen.SignUp, clearBackStack = true)
-        linkAccountManager.logout()
-    }
-
     fun onFocusRequested() {
         updateViewState {
             it.copy(requestFocus = false)
